fixes for fish materials

+++ Assets/Scripts/Fish/Fish.cs
@@ -24,9 +24,10 @@ public class Fish : MonoBehaviour
         mySpriteRenderer.gameObject.SetActive(true);
 
         mySpriteRenderer.material.mainTexture = data.albumSprite;
-        var sizeX = data.albumSprite.width / data.pixelPerMeter;
-        var sizeY = data.albumSprite.height / data.pixelPerMeter;
+        var sizeX = (float)data.albumSprite.width / (float)data.pixelPerMeter;
+        var sizeY = (float)data.albumSprite.height / (float)data.pixelPerMeter;
         mySpriteRenderer.transform.localScale = new Vector3(sizeX, sizeY, 1);
+        
         var box = GetComponent<BoxCollider>();
         box.size = new Vector3(sizeX, sizeY, 1);
     }
